# swagger.yaml
openapi: 3.0.0
info:
  title: MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/
    description: Localhost development server


tags:
- name: Get Routes
  description: Routes that get products

components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: auth-token

  schemas:
    city:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        KnownFor:
          type: String
        population:
          type: Number
        Capital:
          type: Boolean
    
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        date:
          type: string

paths:

  /user/register:
    post:
      summary: Register a new User
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      
      responses:
        200:
          description: User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string

  /city:
    get:
      tags:
      - Get Routes
      summary: Retrieves a list of citys
      description: Retrieves a list of citys as JSON objects.
      responses:
        200:
          description: A list of citys JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/city"
    
    post:
      summary: Create a new city
      description: Create a new city
      security:
        - ApiKeyAuth: []      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/city"

      responses:
        201:
          description: city created succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/city"

  /city/name/{name}:
    get:
      tags:
      - Get Routes
      summary: Specific city by name
      description: Retrieves a specific city based on it name.
      parameters:
        - in: path
          name: name
          required: true
          description: MongoDB name
          schema:
            type: string
      responses:
        200:
          description: A city in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/city"
  
  /city/{id}:
    get:
      tags:
      - Get Routes
      summary: Specific city by Id
      description: Retrieves a specific city based on it id.
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      responses:
        200:
          description: A city in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/city"
   
    put:
      summary: Updates a specific city
      description: Updates a specific city based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/city"

      responses:
        201:
          description: city updated succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/city"

    delete:
      summary: Deletes a specific city
      description: Deletes a specific city based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string      

      responses:
        201:
          description: city deleted succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/city"

    